app/backend/requirements.txt
Django==5.2
djangorestframework==3.14.0
mysqlclient==2.1.0       # For MySQL database connection
psycopg2-binary>=2.9

